Serve our static pages from server.
