--- src/html/xlsxToPdf.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <link rel="stylesheet" href="xlsxToPdf_style.css">
-        <meta charset="UTF-8">
-        <title>External Converter</title>
-        <link rel="preconnect" href="https://fonts.googleapis.com">
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-        <link href="https://fonts.googleapis.com/css2?family=Koulen&display=swap" rel="stylesheet">
-    </head>
-    <body>
-        <div class="mainPage"></div>
-        <div class="rectangle">
-            <div class="container2">
-                <div class="text">
-                    <h1>External Converter to .pdf</h1>
-                </div>
-            </div>
-        </div>
-        <div class="container">
-            <p>.xlsx to .pdf</p>
-            <div class="container1">
-                <div class="icons">
-                    <div class="convert1">
-                        <img src="xlsx2.png" width="100px" height="100px">
-                    </div>
-                    <div class="arrow">
-                        <img src="arrow.png" width="100px" height="100px">
-                    </div>
-                    <div class="convert2" >
-                        <img src="pdf.png" width="100px" height="100px">
-                    </div>
-                </div>
-                <div class="card">
-                    <input type="file" id="file">
-                    <input type="submit" id="btn_primary" value="Convert">
-                </div>
-            </div>
-        </div>
-        <script src="xlsxToPdf_script.js"></script>
-    </body>
-</html>
