--- src/html/xlsxToPdf_style.css
@@ -1,130 +1,0 @@
-body{
-    font-family: 'Koulen', cursive;
-    margin: 0;
-    color: #ffffff;
-    background-color: #AED6DC;
-}
-
-.mainPage {
-    width: 100%;
-}
-
-.rectangle{
-    width: 100%;
-    height: 100px;
-    background-color: #4A536B;
-    border-radius: 0 0 10px 10px;
-}
-
-.container{
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 100px;
-    width: 700px;
-    height: 333px;
-    border-radius: 10px;
-    background-color: #FF9A8D;
-    text-align: center;
-    font-size: 25px;
-}
-
-.container1{
-    width: 100%;
-    max-width: 350px;
-    margin: 0 auto;
-    display: flex;
-    justify-content: center;
-}
-
-.icons {
-    display: flex;
-    width: 100%;    
-    justify-content: space-between;
-    margin-top: 20px;
-}
-
-.arrow{
-    margin-right: 10px;
-}
-
-.buttons{
-    background-color: #4A536B; 
-    border-radius: 10px;
-    height: 34px;
-    width: 130px;
-    text-decoration: none;
-    color: #ffffff;
-    margin-top: 130px;
-    text-align: center;
-    font-size: 20px;
-}
-
-.container2{
-    width: 100%;
-    max-width: 380px;
-    margin: 0 auto;
-    text-align: center;
-}
-
-.text{
-    font-size: 18px;
-    position:absolute;
-    margin-top: auto;
-}
-
-.files{
-    max-width: 500px;
-    position:absolute;
-    display: flex;
-    justify-content: center;
-    margin-top: 150px;
-    margin-left: 15px;
-}
-
-.arrow{
-    width: 100px;
-    height: 100px;
-    color: #4A536B;
-}
-
-input[type=file]{
-    display: none;
-}
-
-.card {
-    width: 100%;
-    max-width: 250px;
-    margin: 0 auto;
-    position: absolute;
-    display: flex;
-    justify-content: space-between;
-}
-
-.btn{
-    width: 120px;
-    color: #fff;
-    border-radius: 99px;
-    border: 1px solid #4A536B;
-    font-family: 'Koulen', cursive;
-    padding: 0.5rem 1.5rem;
-    outline: none;
-    background: #4A536B;
-    transition: all 0.22s;
-    cursor: pointer;
-    margin-top: 180px;
-}
-
-#btn_primary{
-    width: 120px;
-    height: 41px;
-    color: #fff;
-    border-radius: 99px;
-    border: 1px solid #4A536B;
-    font-family: 'Koulen', cursive;
-    padding: 0.5rem 1.5rem;
-    outline: none;
-    background: #4A536B;
-    transition: all 0.22s;
-    cursor: pointer;
-    margin-top: 180px;
-}
